Fixing minor bugs in specific Test-Kit components

+++ sip-test-kit/src/main/java/de/ikor/sip/foundation/testkit/workflow/whenphase/routeinvoker/impl/FtpRouteInvoker.java
@@ -113,7 +113,7 @@ public class FtpRouteInvoker implements RouteInvoker {
 
   private void prepareOtherHeaders(Map<String, Object> headers) {
     if (headers.containsKey(FILE_LAST_MODIFIED)) {
-      Long lastModifiedTimestamp = (Long) headers.get(FILE_LAST_MODIFIED);
+      Long lastModifiedTimestamp = Long.valueOf((Integer) headers.get(FILE_LAST_MODIFIED));
       headers.putIfAbsent(MESSAGE_TIMESTAMP, lastModifiedTimestamp);
     }
   }
